Multiple excluded folders/patterns instead only one
Previously the plugin checked if a folder matches a folder path or pattern and stopped searching for other patterns/excluded folder after its first find.
This meant you couldn't "override" an already excluded folder in the settings. So if you had a regex that matches every folder it would just use the settings you set for the regex and ignore everything else that would also match the folder.